Update primitive variable

+++ ProjectBookingFlights/src/main/java/com/airport/ProjectBookingFlights/repositories/IClientRepository.java
@@ -13,6 +13,6 @@ public interface IClientRepository extends JpaRepository<Client, Long>{
     
     // Clientes que han realizado más de N reservas
     @Query("SELECT c FROM Client c WHERE SIZE(c.reservations) > :minReservations")
-    Set<Client> findClientsWithMoreThanNReservations(int minReservations);
+    Set<Client> findClientsWithMoreThanNReservations(Long minReservations);
 
 }
